+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## [0.1.0] - 2022-10-07
 
-- Initial release
+The initial release, offering feature parity with API's custom classes.
+
+- Supply DummyGTIN, OldDummyGTIN and FallbackDummyGTIN as custom classes
+- Added tests for all core BarcodeValidation GTIN classes and our own custom ones
+- Added tests to prove GTINs with invalid check digit still parse as the correct class